chore(MoveFilesofComponent):move files of steam component [COM-62]
