+++ src/components/Navbar/Navbar.css
@@ -1,5 +1,6 @@
 #container {
-    background-color: #0F6FE4; 
+    background-color: #0F6FE4;
+     
 }
 
 #container-navbar img {
