+++ src/components/NavbarB/NavbarB.css
@@ -1,20 +1,20 @@
-#container {
+.container {
     background-color: #2e2e2d; 
 }
 
-#container-navbar img {
+.logo-startup{
     width: 60px;
     height: 60px;
 }
 
-#text-dropdown {
+.text-dropdown {
     color: #FFFFFF; 
 }
 
-#text-dropdown:hover {
+.text-dropdown:hover {
     background-color: #0F6FE4; 
 }
 
-#text {
+.text {
     color: #FFFFFF; 
 }
